Prevent blobs from being squeezed out of arena
I noticed the scoop could be used to vanish blobs from the arena - now
their position is constrained if detected outside.

Have made it so this function takes a NumberBlob as an argument -
possibly making it more general. Will do the same for other functions
as part of a “re-factoring” exercise.